pyproject.toml
[tool.poetry]
name = "mert"
version = "0.1.0"
description = ""
authors = ["AdmiralDesu <skyrim.men024@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
asyncpg = "^0.25.0"
yt-dlp = "^2022.4.8"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.36"}
alembic = "^1.7.7"
celery = "^5.2.6"
sqlmodel = "^0.0.6"
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
passlib = "^1.7.4"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
